Split classes into 2 files
