+++ parameters.py
@@ -1,29 +1,3 @@
-""" These are the two mains classes. """
-
-import random as rd
-
-
-class attributes:
-    
-    def __init__(self, attribute, name):
-        self.attribute = attribute
-        self.length = len(attribute)
-        self.name = name
-        return 
-        
-    def clear(self):
-        self.attribute = []
-        return self.attribute
-    
-    def add_attr(self, attribute):
-        self.attribute.append(attribute)
-        return self.attribute
-    
-    def pick(self):
-        attr = rd.choice(self.attribute)
-        return attr
-    
-    
 class parameters:
     
     def __init__(self, param):
